refactor(notifications): move components into common folder
