+++ src/components/notifications/mark-all-as-read-button.tsx
